factor out reading reference catalogs to reference.py
